Subgraph: Mappings moved into their own folder
