+++ subgraph/src/mappings/Hatch.ts
@@ -4,14 +4,13 @@ import {
   Close as CloseEvent,
   Contribute as ContributeEvent,
   Refund as RefundEvent,
-} from '../generated/templates/Hatch/Hatch'
+} from '../../generated/templates/Hatch/Hatch'
 import {
   getHatchConfigEntity,
   getContributionEntity,
-  getContributorEntity,
   getHatchState,
-} from './helpers'
-import { getIntStateByKey, getStateByKey } from './hatch-states'
+} from '../utils'
+import { getIntStateByKey, getStateByKey } from '../hatch-states'
 
 export function handleSetOpenDate(event: SetOpenDateEvent): void {
   const config = getHatchConfigEntity(event.address)
@@ -48,12 +47,8 @@ export function handleContribute(event: ContributeEvent): void {
     params.contributor,
     params.vestedPurchaseId
   )
-  const contributor = getContributorEntity(event.address, params.contributor)
   const config = getHatchConfigEntity(event.address)
   const stateKey = getHatchState(event.address)
-
-  contributor.totalValue = contributor.totalValue.plus(params.value)
-  contributor.totalAmount = contributor.totalAmount.plus(params.amount)
 
   config.totalRaised = config.totalRaised.plus(params.value)
   config.stateInt = getIntStateByKey(stateKey)
@@ -74,14 +69,12 @@ export function handleContribute(event: ContributeEvent): void {
   )
 
   config.save()
-  contributor.save()
   contribution.save()
 }
 
 export function handleRefund(event: RefundEvent): void {
   const params = event.params
   const config = getHatchConfigEntity(event.address)
-  const contributor = getContributorEntity(event.address, params.contributor)
   const contribution = getContributionEntity(
     event.address,
     params.contributor,
@@ -91,8 +84,6 @@ export function handleRefund(event: RefundEvent): void {
 
   config.state = getStateByKey(stateKey)
   config.stateInt = getIntStateByKey(stateKey)
-  contributor.totalValue = contributor.totalValue.minus(params.value)
-  contributor.totalAmount = contributor.totalAmount.minus(params.amount)
 
   log.debug(
     'Refund event received. contributor: {} value: {} amount: {} vestedPurchaseId: {}',
@@ -105,6 +96,5 @@ export function handleRefund(event: RefundEvent): void {
   )
 
   config.save()
-  contributor.save()
   store.remove('Contribution', contribution.id)
 }
