+++ subgraph/src/mappings/HatchOracle.ts
@@ -2,8 +2,8 @@ import { store, log } from '@graphprotocol/graph-ts'
 import {
   ScoreTokenSet as ScoreTokenSetEvent,
   RatioSet as RatioSetEvent,
-} from '../generated/templates/HatchOracle/HatchOracle'
-import { getHatchOracleConfigEntity, loadTokenData } from './helpers'
+} from '../../generated/templates/HatchOracle/HatchOracle'
+import { getHatchOracleConfigEntity, populateToken } from '../utils'
 
 export function handleScoreTokenSet(event: ScoreTokenSetEvent): void {
   const params = event.params
@@ -14,7 +14,7 @@ export function handleScoreTokenSet(event: ScoreTokenSetEvent): void {
     event.params.score.toHexString(),
   ])
 
-  const success = loadTokenData(params.score)
+  const success = populateToken(params.score)
 
   if (success) {
     hatchOracle.scoreToken = params.score.toHexString()
